Add installPlugin support for single plugin files (#2033)
## Motivation for the change, related issues

When testing #2029, I wanted to be able to install a plugin from a
single PHP file and have it activated as part of the same step. This
seemed reasonable, so I added support.

## Implementation details

Prior to this PR, if the `installPlugin` step received a file resource
it assumed it was a Zip file. This PR generally continues that practice
but makes an exception for file names ending with a `.php` extension.
File with names ending with `.php` are simply written to
`wp-content/plugins/<filename>`.

## Testing Instructions (or ideally a Blueprint)

CI - this PR adds a relevant unit test