Add deploy workflow for standalone CORS proxy (#2022)
## Motivation for the change, related issues

It is likely that some folks will attempt to abuse the CORS proxy, and
if they succeed despite our defensive measures, we would like to limit
the risk of that abuse being associated with the Playground web app
itself. Let's separate the CORS proxy from the web app.

## Implementation details

This PR adds a workflow to deploy the CORS proxy to a dedicated host.

We need to help folks transition from the CORS proxy on
playground.wordpress.net to the proxy on the dedicated host. We could
even consider keeping a more restricted version of the proxy on
playground.wordpress.net, though having two instances would likely be
confusing.

## Testing Instructions (or ideally a Blueprint)

- After merge, manually run the deployment workflow and confirm that the CORS proxy
works on the dedicated host.