Ensure that Site Editor templates are associated with the correct taxonomy (#1997)
Fixes #1996 

While the bug might be in the WordPress-Importer itself, we can fix it
by adding a filter that just ensures that the term will be created. I
vote for adding this small intermediate fix even as in #1888 there is
the plan to import WXRs differently in future. This is confusing
behavior right now and we can fix it