[Website] Don't show the error reporting modal on the initial load (#2068)
Hides the error report modal when the page is loaded with
`?modal=error-report`.

During the development I often run into an error and then refresh the
page without closing the modal. On page reload, I see the error
reporting modal again. That's not useful as it's not populated with any
content. Therefore, this PR discards the `modal=` query argument if it's
set to `error-report` on the initial page load.

 ## Testing instructions

Go to http://127.0.0.1:5400/website-server/?modal=error-report and
confirm you don't see the error reporting modal