[Blueprints] Preserve the first char of all filenames sourced from GitDirectoryReference (#2070)
Adjust GitDirectoryReference to stop cutting off the first character
from each filename when the repository path starts with a slash.

Closes https://github.com/WordPress/wordpress-playground/issues/2061
Closes #1999

 ## Implementation

The problem was passing the wring prefix length to
`fileName.substring()` (user input vs path without the leading slash).
This PR adjusts that.

In addition, it updates git-sparse-checkout.ts to use `commitHash` and
not ref name for fetching files. This way you can checkout a specific
commit, and that's convenient for testing and developers consuming the
API.

It also modified the CORS proxy to allow 'git-protocol' in headers.

## Testing 

Go to
http://127.0.0.1:5400/website-server/#{%22landingPage%22:%22/wp-admin/%22,%22preferredVersions%22:{%22php%22:%227.4%22,%22wp%22:%225.9%22},%22steps%22:[{%22step%22:%22login%22,%22username%22:%22admin%22},{%22step%22:%22installPlugin%22,%22pluginData%22:{%22resource%22:%22git:directory%22,%22url%22:%22https://wordpress-playground-cors-proxy.net/cors-proxy.php?https://github.com/Automattic/page-optimize.git%22,%22ref%22:%22master%22,%22path%22:%22/%22},%22options%22:{%22activate%22:true,%22targetFolderName%22:%22page-optimize-from-git%22}}]}
and confirm you see wp-admin (without CSS).