Use transparent CORS proxy in Blueprint Builder (#2089)
## Motivation for the change, related issues

The Blueprint builder doesn't use the CORS proxy transparently like the
web app does. The reason is that the builder does not pass a CORS proxy
URL to `startPlaygroundWeb()`.

## Implementation details

This PR does a liberal conversion of builder.js to builder.ts and
imports the virtual vite module `virtual:cors-proxy-url` to obtain the
right CORS proxy URL depending on target.

## Testing Instructions (or ideally a Blueprint)

- `npm run build`
- Serve `dist/packages/playground/wasm-wordpress-net` via a local web
server
- Navigate to the Blueprint builder on the server
- Run the following Blueprint and confirm that the Nautilus theme is
installed and activated:
```json
{
  "landingPage": "/wp-admin/themes.php",
  "steps": [
    {
      "step": "login",
      "username": "admin"
    },
    {
      "step": "installTheme",
      "themeZipFile": {
        "resource": "url",
        "url": "https://codeload.github.com/ndiego/nautilus/zip/refs/heads/main"
      }
    }
  ]
}
```
- Play with Blueprint Builder and confirm that all appears to be in
working order.